AppContext moved to shared partition - CoreRT specific changes
