+++ src/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -190,7 +190,7 @@
     <Compile Include="System\Text\CodePageDataItem.cs" />
     <Compile Include="System\__Canon.cs" Condition="'$(IsProjectNLibrary)' == 'true'" />
     <Compile Include="System\Activator.cs" />
-    <Compile Include="System\AppContext.cs" />
+    <Compile Include="System\AppContext.CoreRT.cs" />
     <Compile Include="System\AppContextConfigHelper.cs" />
     <Compile Include="System\Buffer.cs" />
     <Compile Include="System\Collections\Generic\ArraySortHelper.CoreRT.cs" />
@@ -335,7 +335,7 @@
     <Compile Include="Internal\Runtime\Augments\RuntimeThread.Windows.cs" />
     <Compile Include="Microsoft\Win32\SafeHandles\SafeThreadPoolIOHandle.cs" />
     <Compile Condition="'$(EnableWinRT)' == 'true'" Include="System\TimeZoneInfo.WinRT.cs" />
-    <Compile Include="System\AppContext.Windows.cs" />
+    <Compile Include="System\AppContext.Windows.CoreRT.cs" />
     <Compile Include="System\DateTime.Windows.CoreRT.cs" />
     <Compile Include="System\DateTime.Win32.cs" Condition="'$(EnableWinRT)'!='true'"/>
     <Compile Include="System\DateTime.WinRT.cs" Condition="'$(EnableWinRT)'=='true'"/>
@@ -498,7 +498,7 @@
   <ItemGroup Condition="'$(TargetsUnix)'=='true'">
     <Compile Include="Internal\Runtime\Augments\RuntimeThread.Unix.cs" />
     <Compile Include="Microsoft\Win32\SafeHandles\SafeWaitHandle.Unix.cs" />
-    <Compile Include="System\AppContext.Unix.cs" />
+    <Compile Include="System\AppContext.Unix.CoreRT.cs" />
     <Compile Include="System\DateTime.Unix.CoreRT.cs" />
     <Compile Include="System\Environment.Unix.cs" />
     <Compile Include="System\HighPerformanceCounter.Unix.cs" />
