+++ src/System.Private.CoreLib/src/System/AppContext.Windows.CoreRT.cs
